requirements.txt
albumentations==1.3.0
anyio==3.6.2
arpa==0.1.0b4
boto3==1.26.131
botocore==1.29.131
click==8.1.3
colorama==0.4.6
fastapi==0.95.1
filelock==3.12.0
h11==0.14.0
idna==3.4
imageio==2.28.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
lazy_loader==0.2
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
opencv-python-headless==4.7.0.72
packaging==23.1
Pillow==9.5.0
pydantic==1.10.7
python-dateutil==2.8.2
python-dotenv==1.0.0
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
s3transfer==0.6.1
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
sympy==1.11.1
threadpoolctl==3.1.0
tifffile==2023.4.12
torch==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.22.0
